Move announce participant code to service
